Fixed getting all users including those not cached at first. Also added another property (usesDB) to the commands.
Used message.guild.members.fetch(). I guess using the client only gets the cache?

I added the property, usesDB, because previously bot.js would connect to the database every time a command was executed which probably took up a lot of time and now it only connects if the command uses the database.